fix: split model chart and switch to tps
