+++ src/app/_components/browse/ModelPreviewCard.tsx
@@ -45,7 +45,7 @@ export const ModelPreviewCard = ({
           {/* Header */}
           <div className="flex items-start justify-between">
             <div className="flex flex-col">
-              <h3 className="font-medium text-sm text-gray-900 transition-colors group-hover:text-gray-700">
+              <h3 className="text-sm font-medium text-gray-900 transition-colors group-hover:text-gray-700">
                 {name?.split("/")[1]}
               </h3>
               <p className="text-xs text-gray-500">{name?.split("/")[0]}</p>
